refactor(cta): 👷 move cta banners up
